Move transaction types to transaction utilities
