+++ src/components/TransactionalInfo.tsx
@@ -22,8 +22,8 @@ import { useTranslation } from 'react-i18next'
 import { useParams } from 'react-router-dom'
 import styled, { useTheme } from 'styled-components'
 
-import { AddressHash, SimpleTx, TransactionDirection, useAddressesContext } from '../contexts/addresses'
-import { isExplorerTransaction, isSimpleTx } from '../utils/transactions'
+import { AddressHash, SimpleTx, useAddressesContext } from '../contexts/addresses'
+import { isExplorerTransaction, isSimpleTx, TransactionDirection } from '../utils/transactions'
 import AddressBadge from './AddressBadge'
 import Amount from './Amount'
 import Badge from './Badge'
